Extract MD5 checksum validation to a standalone function
